move SKPaymentTransation tests to another file
